+++ TOMB5/game/debris.h
@@ -7,4 +7,4 @@ void TriggerDebris(GAME_VECTOR* pos, void* TextInfo, short* Offsets, long* Vels,
 long GetFreeDebris();
 void UpdateDebris();
 
-#define ShatterObject	( (void(__cdecl*)(SHATTER_ITEM*, MESH_INFO*, short, short, int)) 0x0041D6B0 )
+#define ShatterObject	( (void(__cdecl*)(SHATTER_ITEM*, MESH_INFO*, short, short, long)) 0x0041D6B0 )
