+++ TOMB5/specific/alexstuff.h
@@ -8,5 +8,5 @@ void aFixUpRoom(ROOM_INFO* r, char* s);
 void aUpdate();
 void aInitWater();
 
-#define aFetchCutData	( (char*(__cdecl*)(int)) 0x00491F60 )
-#define DoCredits	( (int(__cdecl*)()) 0x004927C0 )
+#define aFetchCutData	( (char*(__cdecl*)(long)) 0x00491F60 )
+#define DoCredits	( (long(__cdecl*)()) 0x004927C0 )
